Set up unittest exposition repository
This commit will add a unittest version of one of our bot's test
files: `test_converters.py`. It also includes the old version of
the same file to make a comparison easier.

It also includes a very minimal `Pipfile` and `Pipfile.lock` to
set up the required environment for running the tests. To make
the style consistent, I've included the same `dev` packages and
`flake8` configuration file as we actually use in the bot.

To compare both frameworks, you can run them with:

- `pipenv run unittest`

- `pipenv run pytest`